+++ GrailDiary/NotebookStructureViewController.swift
@@ -164,6 +164,15 @@ final class NotebookStructureViewController: UIViewController {
     return view
   }()
 
+  private let footerRegistration = UICollectionView.SupplementaryRegistration<UICollectionViewListCell>(elementKind: UICollectionView.elementKindSectionFooter) { footerView, _, _ in
+    var footerConfiguration = footerView.defaultContentConfiguration()
+    footerConfiguration.text = "Version \(UIApplication.versionString)"
+    footerConfiguration.textProperties.font = UIFont.preferredFont(forTextStyle: .caption1)
+    footerConfiguration.textProperties.color = UIColor.secondaryLabel
+    footerView.contentConfiguration = footerConfiguration
+  }
+
+
   private lazy var dataSource: UICollectionViewDiffableDataSource<Section, Item> = {
     let hashtagRegistration = UICollectionView.CellRegistration<UICollectionViewListCell, Item> { cell, _, item in
       var contentConfiguration = cell.defaultContentConfiguration()
@@ -184,15 +193,8 @@ final class NotebookStructureViewController: UIViewController {
       view.dequeueConfiguredReusableCell(using: hashtagRegistration, for: indexPath, item: item)
     }
 
-    dataSource.supplementaryViewProvider = { [weak dataSource] collectionView, kind, indexPath in
+    dataSource.supplementaryViewProvider = { [footerRegistration] collectionView, kind, indexPath in
       if kind == UICollectionView.elementKindSectionFooter {
-        let footerRegistration = UICollectionView.SupplementaryRegistration<UICollectionViewListCell>(elementKind: UICollectionView.elementKindSectionFooter) { footerView, _, _ in
-          var footerConfiguration = footerView.defaultContentConfiguration()
-          footerConfiguration.text = "Version \(UIApplication.versionString)"
-          footerConfiguration.textProperties.font = UIFont.preferredFont(forTextStyle: .caption1)
-          footerConfiguration.textProperties.color = UIColor.secondaryLabel
-          footerView.contentConfiguration = footerConfiguration
-        }
         return collectionView.dequeueConfiguredReusableSupplementary(using: footerRegistration, for: indexPath)
       }
       Logger.shared.error("Unexpected supplementary kind \(kind), returning nil")